+++ src/test/kotlin/uiautotests/model/AjaxDataPage.kt
@@ -3,16 +3,19 @@ package uiautotests.model
 import com.codeborne.selenide.*
 import com.codeborne.selenide.Selenide.element
 import org.openqa.selenium.*
+import java.time.Duration
+import kotlin.test.assertEquals
 
 class AjaxDataPage {
     private val triggeringAjaxRequestButton = By.id("ajaxButton")
     private val fieldOfLoadedData = By.className("bg-success")
 
     fun clickTriggeringAjaxRequestButton() {
-        Selenide.element(triggeringAjaxRequestButton).should(Condition.visible).click()
+        element(triggeringAjaxRequestButton).should(Condition.visible).click()
     }
 
-    fun checkLoadedDataInTheField(value: String): SelenideElement {
-       return element(fieldOfLoadedData).shouldHave(Condition.text(value))
+    fun checkLoadedDataInTheField(value: String) {
+        element(fieldOfLoadedData).shouldBe(Condition.visible, Duration.ofSeconds(16))
+        assertEquals((element(fieldOfLoadedData).text), value)
     }
 }
