+++ src/test/kotlin/uiautotests/model/LoadDelayPage.kt
@@ -1,13 +1,15 @@
 package uiautotests.model
 
 import com.codeborne.selenide.Condition
-import com.codeborne.selenide.Selenide
+import com.codeborne.selenide.Selenide.element
 import org.openqa.selenium.By
+import java.time.Duration
 
 class LoadDelayPage {
-    private val appearingAfterDelayButton = By.xpath("//button[contains(concat(' ', normalize-space(@class), ' '), ' btn-primary ')]")
+    private val appearingAfterDelayButton =
+        By.xpath("//button[contains(concat(' ', normalize-space(@class), ' '), ' btn-primary ')]")
 
     fun clickAppearingAfterDelayButton() {
-        Selenide.element(appearingAfterDelayButton).should(Condition.visible).click()
+        element(appearingAfterDelayButton).shouldBe(Condition.visible, Duration.ofSeconds(15)).click()
     }
 }
