+++ src/test/kotlin/uiautotests/uitestingplayground/loaddelay/LoadDelayTest.kt
@@ -26,9 +26,6 @@ class LoadDelayTest: BaseTest() {
     @DisplayName("Ожидание загрузки кнопки для клика осуществляется")
     fun clickAppearingAfterDelayButton() {
         mainPageForUiTests.goToLoadDelayPage()
-        /* Элемент ожидается из-за установки глобального таймаута
-           Configuration.timeout = 20000
-           в классе BaseTest */
         loadDelayPage.clickAppearingAfterDelayButton()
     }
 }
